+++ src/esp32/src/packets/packets.h
@@ -27,7 +27,7 @@ uint8_t *write_register_request_normalized(uint16_t *packetLength);
 uint8_t *write_auth_request_normalized(uint16_t *packetLength);
 uint8_t *write_revoke_auth_request_normalized(uint16_t *packetLength);
 uint8_t *write_data_packet_normalized(uint8_t *data, uint32_t length, uint16_t *packetLength);
-uint8_t *write_data_value_packet_normalized(uint8_t dataType, uint64_t timeStamp, double data, uint16_t *packetLength);
+uint8_t *write_data_value_packet_normalized(uint8_t dataType, uint64_t timeStamp, float data, uint16_t *packetLength);
 uint8_t *write_device_status_packet_normalized(uint64_t timeStamp, uint32_t status, uint16_t *packetLength);
 uint8_t *write_log_packet_normalized(uint64_t timeStamp, char *level, uint32_t levelLength, char *message, uint32_t messageLength, uint16_t *packetLength);
 #endif
